:fire: Remove comma detection for last register
On the last register seen, it should not matter that the next parsed
token is valid. Imagine the following situation with the old rule of
raising an exception if a comma or another token shows up after a group
of registers.

ADD R3, R4, R4

This sets R3 = R4 + R4. In this case, the rule is OK. However...

ADD R3, R4, #5

This sets R3 = R4 + int(5). In this case, the register group R3, R4 has
a comma and additional sequences afterwards. Hence, this change is made.