:rocket: Finish Parser
Finished the parser, aside from some lingering TODO comments that will
be knocked out in a later iteration. The `input.asm` is an example of
all the instructions that currently work. To be added to the parser:
directives...

Most importantly, `.orig` and `.end` to be able to begin/end blocks
of assembly. The assembler, for now, will ignore the existence of
`.orig` statements and simply assemble each line into binary.

To support binary file creation, these directives will need to be
implemented.